BW-1354 - Porting CBAS preliminary step (#6837)
* Getting rid of shared utility file; Adding/Updating WES version of submit.

* Edit spec file

* Adding Wes-like error
