+++ engine/src/main/scala/cromwell/webservice/routes/wes/WesRouteSupport.scala
@@ -1,41 +1,47 @@
 package cromwell.webservice.routes.wes
 
 import akka.actor.ActorRef
-import akka.http.scaladsl.model.{StatusCode, StatusCodes}
+import akka.http.scaladsl.model.{Multipart, StatusCode, StatusCodes}
 import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.server.directives.RouteDirectives.complete
 import akka.http.scaladsl.server.{Directive1, Route}
 import akka.pattern.{AskTimeoutException, ask}
+import akka.stream.ActorMaterializer
 import akka.util.Timeout
+import cats.data.NonEmptyList
 import com.typesafe.config.ConfigFactory
-import cromwell.core.WorkflowId
 import cromwell.core.abort.SuccessfulAbortResponse
+import cromwell.core.{WorkflowId, WorkflowOnHold, WorkflowState, WorkflowSubmitted}
 import cromwell.engine.instrumentation.HttpInstrumentation
 import cromwell.engine.workflow.WorkflowManagerActor.WorkflowNotFoundException
+import cromwell.engine.workflow.workflowstore.{WorkflowStoreActor, WorkflowStoreSubmitActor}
 import cromwell.server.CromwellShutdown
 import cromwell.services.metadata.MetadataService.{BuildMetadataJsonAction, GetSingleWorkflowMetadataAction, GetStatus, MetadataServiceResponse, StatusLookupFailed}
 import cromwell.services.{FailedMetadataJsonResponse, SuccessfulMetadataJsonResponse}
-import cromwell.webservice.WebServiceUtils.EnhancedThrowable
+import cromwell.webservice.PartialWorkflowSources
+import cromwell.webservice.WebServiceUtils.{EnhancedThrowable, completeResponse, materializeFormData}
+import cromwell.webservice.routes.CromwellApiService
 import cromwell.webservice.routes.CromwellApiService.{UnrecognizedWorkflowException, validateWorkflowIdInMetadata}
 import cromwell.webservice.routes.MetadataRouteSupport.{metadataBuilderActorRequest, metadataQueryRequest}
 import cromwell.webservice.routes.wes.WesResponseJsonSupport._
-import cromwell.webservice.routes.wes.WesRouteSupport._
-import cromwell.webservice.routes.{CromwellApiService, WesCromwellRouteSupport}
+import cromwell.webservice.routes.wes.WesRouteSupport.{respondWithWesError, _}
 import net.ceedubs.ficus.Ficus._
 
 import scala.concurrent.duration.FiniteDuration
-import scala.concurrent.{ExecutionContext, Future}
+import scala.concurrent.{ExecutionContext, Future, TimeoutException}
 import scala.util.{Failure, Success}
 
 
 
-trait WesRouteSupport extends HttpInstrumentation with WesCromwellRouteSupport {
+trait WesRouteSupport extends HttpInstrumentation {
 
   val serviceRegistryActor: ActorRef
   val workflowManagerActor: ActorRef
+  val workflowStoreActor: ActorRef
 
   implicit val ec: ExecutionContext
   implicit val timeout: Timeout
+  implicit val materializer: ActorMaterializer
 
   /*
     Defines routes intended to sit alongside the primary Cromwell REST endpoints. For instance, we'll now have:
@@ -68,7 +74,7 @@ trait WesRouteSupport extends HttpInstrumentation with WesCromwellRouteSupport {
               } ~
                 post {
                   extractSubmission() { submission =>
-                    submitRequest(submission.entity,
+                    wesSubmitRequest(submission.entity,
                       isSingleSubmission = true)
                   }
                 }
@@ -108,6 +114,62 @@ trait WesRouteSupport extends HttpInstrumentation with WesCromwellRouteSupport {
         }
       )
     }
+
+  def toWesResponse(workflowId: WorkflowId, workflowState: WorkflowState):  WesRunStatus = {
+    WesRunStatus(workflowId.toString, WesState.fromCromwellStatus(workflowState))
+  }
+
+  def toWesResponseId(workflowId: WorkflowId): WesRunId ={
+    WesRunId(workflowId.toString)
+  }
+
+  def wesSubmitRequest(formData: Multipart.FormData, isSingleSubmission: Boolean): Route = {
+    def getWorkflowState(workflowOnHold: Boolean): WorkflowState = {
+      if (workflowOnHold)
+        WorkflowOnHold
+      else WorkflowSubmitted
+    }
+
+    def sendToWorkflowStore(command: WorkflowStoreActor.WorkflowStoreActorSubmitCommand, warnings: Seq[String], workflowState: WorkflowState): Route = {
+      // NOTE: Do not blindly copy the akka-http -to- ask-actor pattern below without knowing the pros and cons.
+      onComplete(workflowStoreActor.ask(command).mapTo[WorkflowStoreSubmitActor.WorkflowStoreSubmitActorResponse]) {
+        case Success(w) =>
+          w match {
+            case WorkflowStoreSubmitActor.WorkflowSubmittedToStore(workflowId, _) =>
+              completeResponse(StatusCodes.Created, toWesResponseId(workflowId), warnings)
+            case WorkflowStoreSubmitActor.WorkflowsBatchSubmittedToStore(workflowIds, _) =>
+              completeResponse(StatusCodes.Created, workflowIds.toList.map(toWesResponse(_, workflowState)), warnings)
+            case WorkflowStoreSubmitActor.WorkflowSubmitFailed(throwable) =>
+              respondWithWesError(throwable.getLocalizedMessage, StatusCodes.BadRequest)
+          }
+        case Failure(_: AskTimeoutException) if CromwellShutdown.shutdownInProgress() => respondWithWesError("Cromwell service is shutting down", StatusCodes.InternalServerError)
+        case Failure(e: TimeoutException) => e.failRequest(StatusCodes.ServiceUnavailable)
+        case Failure(e) => e.failRequest(StatusCodes.InternalServerError, warnings)
+      }
+    }
+
+    onComplete(materializeFormData(formData)) {
+      case Success(data) =>
+        PartialWorkflowSources.fromSubmitRoute(data, allowNoInputs = isSingleSubmission) match {
+          case Success(workflowSourceFiles) if isSingleSubmission && workflowSourceFiles.size == 1 =>
+            val warnings = workflowSourceFiles.flatMap(_.warnings)
+            sendToWorkflowStore(WorkflowStoreActor.SubmitWorkflow(workflowSourceFiles.head), warnings, getWorkflowState(workflowSourceFiles.head.workflowOnHold))
+          // Catches the case where someone has gone through the single submission endpoint w/ more than one workflow
+          case Success(workflowSourceFiles) if isSingleSubmission =>
+            val warnings = workflowSourceFiles.flatMap(_.warnings)
+            val e = new IllegalArgumentException("To submit more than one workflow at a time, use the batch endpoint.")
+            e.failRequest(StatusCodes.BadRequest, warnings)
+          case Success(workflowSourceFiles) =>
+            val warnings = workflowSourceFiles.flatMap(_.warnings)
+            sendToWorkflowStore(
+              WorkflowStoreActor.BatchSubmitWorkflows(NonEmptyList.fromListUnsafe(workflowSourceFiles.toList)),
+              warnings, getWorkflowState(workflowSourceFiles.head.workflowOnHold))
+          case Failure(t) => t.failRequest(StatusCodes.BadRequest)
+        }
+      case Failure(e: TimeoutException) => e.failRequest(StatusCodes.ServiceUnavailable)
+      case Failure(e) => respondWithWesError(e.getLocalizedMessage, StatusCodes.InternalServerError)
+    }
+  }
 }
